RedM/FiveM: Increase navigation movement speed.
When holding down directional keys, speed increases with a total of 4 steps depending on the duration of the press. This has been upgraded from only 1 speed increment for cases where menus or lists were (extremely) long. This change is applied to every direction of navigation, including normal menu navigation, as well as (dynamic) list items.